+++ CHANGELOG.md
@@ -11,6 +11,7 @@ project adheres to
 
 * Some improvements in list handling, formatting and introspection.
 * Change default precision to 10, matching dart-sass.
+* Semicolon is optional after variable, if last in block or file.
 
 
 ## Release 0.14.0 - 2020-10-03